requirements.txt
argparse-addons==0.12.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
bitstruct==8.21.0
boto3==1.39.0
botocore==1.39.0
cantools==40.2.3
cffi==1.17.1
channels==4.2.2
confluent-kafka==2.10.1
constantly==23.10.4
crccheck==1.3.0
cryptography==45.0.5
daphne==4.2.1
diskcache==5.6.3
Django==4.2.20
django-cors-headers==4.7.0
djangorestframework==3.16.0
dnspython==2.7.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==1.0.1
numpy==2.3.0
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
python-can==4.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-etcd==0.4.5
pytz==2025.2
s3transfer==0.13.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
textparser==0.24.0
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wrapt==1.17.2
zope.interface==7.2